update: split auth into ipython and core
